requirements.txt
black==25.1.0
click==8.2.1
colorama==0.4.6
deepdiff==8.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy_extensions==1.1.0
orderly-set==5.4.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
Pygments==2.19.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
pygithub